Add files via upload
The Project Hospital Management System (HMS) is for computerizing the working in a hospital. The software takes care of all the requirements of an average hospital and is capable to provide easy and effective storage of information related to patients that come up to the hospital. It generates test reports, provide prescription details including various tests, check-up and medicines prescribed to patient and doctor. It also provides injection details and billing facility. The system also provides the facility of backup as per the requirements.